+++ CHANGELOG.rst
@@ -4,6 +4,22 @@ Changelog
 
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
+1.2.0 (2021-06-09)
+------------------
+
+**Added**
+
+* Added a window for the image viewer (`#28 <https://github.com/qbicsoftware/omero-portlet/pull/28>`_)
+
+**Fixed**
+
+* Cleaned up metadata window
+* Replaced the image size and timepoints columns for a single column
+
+**Dependencies**
+
+**Deprecated**
+
 1.1.0 (2020-10-06)
 ------------------